Split action review projection modules
